+++ core/controllers/agent.js
@@ -13,45 +13,32 @@ module.exports = function(server, passport){
     paramsResolver.customerNameToEntity({}),
     paramsResolver.hostnameToHost({})
   ],controller.update);
-
-  /**
-  return {
-    routes: [
-      {
-        route: '/agent/:hostname',
-        method: 'put',
-        middleware: [
-          paramsResolver.customerNameToEntity({}),
-          paramsResolver.hostnameToHost({}) 
-        ],
-        action: controller.update
-      }
-    ]
-  };
-  */
 }
 
 var controller = {
   update (req, res, next) {
     var host = req.host ;
     if(!host){
-      logger.error('invalid request for id %s. host not found', req.params.hostname);
+      logger.error('invalid request for %s. host not found', req.params.hostname);
       return res.send(404,'invalid host');
     }
 
-    logger.log('receiving agent keep alive for host "%s"',host.hostname);
+    logger.log('receiving agent keep alive for host "%s"', host.hostname);
 
     host.last_update = new Date();
     host.save(err=>{
       if(err) return logger.error(err);
-      let options = {
+
+      ResourceManager.findHostResources(host,{
         'type':'host',
         'ensureOne':true
-      };
-      ResourceManager.findHostResources(host,options,(err,resource)=>{
-        if(err||!resource)return;
+      }, (err,resource) => {
+        if (err||!resource) return;
         var handler = new ResourceManager(resource);
-        handler.handleState({ state:'normal' });
+        handler.handleState({
+          state:'normal',
+          last_update: new Date()
+        });
       });
     });