+++ core/controllers/resource.js
@@ -105,8 +105,7 @@ var controller = {
     }
 
     var manager = new ResourceManager(resource);
-    input.last_update = Date.now();
-
+    input.last_update = new Date();
     manager.handleState(input,function(error){
       if(!error) {
         res.send(200);