+++ core/service/monitor.js
@@ -14,23 +14,23 @@ const Constants = require('../constants/monitors');
 const Scheduler = require('../service/scheduler');
 
 module.exports = {
-  start: function() {
-    let mconfig = config.get('monitor');
+  start: function () {
+    var mconfig = config.get('monitor');
     // to seconds
-    var interval = mconfig.resources_check_failure_interval_milliseconds / 1000;
+    var interval = mconfig.check_interval / 1000;
 
     Scheduler.agenda.define(
       'monitoring',
       { lockLifetime: (5 * 60 * 1000) }, // max lock
       (job, done) => { checkResourcesState(done) }
     );
+
     Scheduler.agenda.every(`${interval} seconds`,'monitoring');
     logger.log('monitoring started');
   }
 };
 
-
-function checkResourcesState(done){
+function checkResourcesState (done) {
   logger.debug('***** CHECKING RESOURCES STATUS *****');
   Resource
   .find({ 'enable': true })
@@ -48,7 +48,7 @@ function checkResourcesState(done){
   });
 }
 
-function runChecks(resource,done) {
+function runChecks (resource,done) {
   CustomerService.getCustomerConfig(
     resource.customer_id,
     function(error,cconfig) {
@@ -56,21 +56,22 @@ function runChecks(resource,done) {
         logger.error('customer %s configuration fetch failed',resource.customer_name);
         return done();
       }
+
       if(!cconfig){
-        logger.error('customer %s configuration not found',resource.customer_name);
+        logger.error('customer %s configuration unavailable',resource.customer_name);
         return done();
       }
 
       switch(resource.type){
         case 'host':
-          checkHostResourceStatus(resource,done);
+          checkHostResourceStatus(resource, cconfig.monitor, done);
           break;
         case 'script':
         case 'scraper':
         case 'process':
         case 'dstat':
         case 'psaux':
-          checkResourceMonitorStatus(resource,cconfig.monitor,done);
+          checkResourceMonitorStatus(resource, cconfig.monitor, done);
           break;
         case 'default':
           logger.error('unhandled resource %s', resource.type);
@@ -92,79 +93,75 @@ function checkResourceMonitorStatus(resource,cconfig,done) {
       logger.error('Resource monitor query error : %s', error.message);
       return done();
     }
+
     if(!monitor){
       logger.debug('resource hasn\'t got any monitor');
       return done();
     }
 
     logger.debug('checking monitor "%s"', resource.name);
-    var last_update = resource.last_update.getTime();
 
-    validLastupdate({
-      'loop_duration': monitor.looptime,
-      'loop_threshold': cconfig.resources_alert_failure_threshold_milliseconds,
-      'last_update': last_update,
-      'fails_count': resource.fails_count
-    },function(error,valid,failedLoops){
-      if(!valid){
-        var manager = new ResourceService(resource);
-        manager.handleState({
-          'state':Constants.RESOURCE_STOPPED,
-          'last_check':Date.now()
-        });
-      }
-      done();
-    });
-  });
-}
-
-function checkHostResourceStatus(resource,done)
-{
-  done||(done=function(){});
-
-  logger.debug('checking host resource %s', resource.name);
-  validLastupdate({
-    'loop_duration':config.get('agent').core_workers.host_ping.looptime,
-    'loop_threshold':config.get('monitor').resources_alert_failure_threshold_milliseconds,
-    'last_update':resource.last_update.getTime(),
-    'fails_count':resource.fails_count
-  },function(error,valid,failedLoops){
-    if(!valid){
+    var trigger = triggerAlert(
+      resource.last_update,
+      monitor.looptime,
+      resource.fails_count,
+      cconfig.fails_count_alert
+    );
+    if( trigger ) {
       var manager = new ResourceService(resource);
       manager.handleState({
-        'state':Constants.RESOURCE_STOPPED,
-        'last_check':Date.now()
+        state: Constants.RESOURCE_STOPPED,
+        last_check: new Date()
       });
+    } else {
+      resource.last_check = new Date();
+      resource.save();
     }
+
     done();
   });
 }
 
-function validLastupdate(options,done)
-{
-  done||(done=()=>{});
-  logger.debug(options);
-  var valid, nowTime = Date.now();
-  var loopDuration = options.loop_duration;
-  var loopThreshold = options.loop_threshold;
-  var failedLoopsCount = options.fails_count;
-  var lastUpdate = options.last_update;
-  var timeElapsed = nowTime - lastUpdate;
-  var updateThreshold = loopDuration + loopThreshold;
+function checkHostResourceStatus (resource, cconfig, done) {
+  done||(done=function(){});
 
-  var elapsedMinutes = Math.floor(timeElapsed/1000/60);
-  logger.debug('last update time elapsed ' + elapsedMinutes + ' minutes');
-
-  var failedLoops = Math.floor(timeElapsed / loopDuration);
-  logger.debug('failed loops count %s', failedLoops);
-  if(timeElapsed > updateThreshold) {
-    if(failedLoops > failedLoopsCount) {
-      logger.debug('last update check failed %s times',failedLoops);
-      done(null,valid=false,failedLoops);
-    } else {
-      done(null,valid=true,failedLoops);
-    }
+  logger.debug('checking host resource %s', resource.name);
+  var agentKeepAliveLoop = config.get('agent').core_workers.host_ping.looptime;
+  var trigger = triggerAlert(
+    resource.last_update,
+    agentKeepAliveLoop,
+    resource.fails_count,
+    cconfig.fails_count_alert
+  );
+  if( trigger ) {
+    var manager = new ResourceService(resource);
+    manager.handleState({
+      state: Constants.RESOURCE_STOPPED,
+      last_check: Date.now()
+    });
   } else {
-    done(null,valid=true);
+    resource.last_check = new Date();
+    resource.save();
   }
+
+  done();
+}
+
+function triggerAlert (lastUpdate,loopDuration,failsCount,failsCountThreshold) {
+  var timeElapsed = Date.now() - lastUpdate.getTime();
+  var loopsElapsed = Math.floor( timeElapsed / loopDuration );
+
+  logger.debug({ 
+    'last update': lastUpdate,
+    'loops elapsed': loopsElapsed,
+    'loop duration': loopDuration,
+    'time elapsed (mins)': (timeElapsed/1000/60)
+  });
+
+  if( loopsElapsed >= 2 ){
+    if( failsCount == 0 ) return true;
+    if( 1 == (loopsElapsed - failsCount) ) return true;
+    if( loopsElapsed > failsCountThreshold ) return true;
+    return false;
+  } else return false;
 }